requirements.txt
openai-whisper
gradio
